[Refator] HeaderProvider
