+++ src/app/(pages)/(auth)/layout.tsx
@@ -1,4 +1,4 @@
-import { AuthPagesLayout } from "@/styles/pageStyle";
+import { AuthPagesLayout } from '@/styles/pageStyle';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return <div className={AuthPagesLayout}>{children}</div>;
