+++ src/components/InitializeAuth.tsx
@@ -7,8 +7,10 @@ import { getUserInfo } from '@/app/api/user';
 
 export default function InitializeAuth() {
   const initializeAuth = useAuthStore((state) => state.initializeAuth);
+  const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
   const setUserInfo = useUserStore((state) => state.setUserInfo);
 
+  // 초기 mount 시
   useEffect(() => {
     const init = async () => {
       initializeAuth();
@@ -30,6 +32,26 @@ export default function InitializeAuth() {
 
     init();
   }, [initializeAuth, setUserInfo]);
+
+  // 로그인 상태가 변경되었을 때도 유저 정보 fetch
+  useEffect(() => {
+    if (!isLoggedIn) return;
+
+    const fetchUser = async () => {
+      try {
+        const user = await getUserInfo();
+        setUserInfo({
+          nickname: user.nickname,
+          profileImage: user.profileImage,
+          teams: user.teams,
+        });
+      } catch (_) {
+        // 유저 정보 요청 실패 무시
+      }
+    };
+
+    fetchUser();
+  }, [isLoggedIn, setUserInfo]);
 
   return null;
 }
