+++ src/stores/useAuthStore.ts
@@ -1,6 +1,7 @@
 'use client';
 
 import { create } from 'zustand';
+import { useUserStore } from './useUserStore';
 
 interface AuthState {
   accessToken: string | null;
@@ -13,14 +14,18 @@ interface AuthState {
 export const useAuthStore = create<AuthState>((set) => ({
   accessToken: null,
   isLoggedIn: false,
+
   setAccessToken: (token: string) => {
     localStorage.setItem('accessToken', token);
     set({ accessToken: token, isLoggedIn: true });
   },
+
   logout: () => {
     localStorage.removeItem('accessToken');
     set({ accessToken: null, isLoggedIn: false });
+    useUserStore.getState().clearUserInfo();
   },
+
   initializeAuth: () => {
     const storedToken = localStorage.getItem('accessToken');
     if (storedToken) {
